+++ src/datastore/sync_methods/ejectAll.js
@@ -3,23 +3,23 @@ var DSErrors = require('../../errors');
 
 function ejectAll(resourceName, params, options) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
+  var definition = _this.defs[resourceName];
   params = params || {};
 
   if (!definition) {
     throw new DSErrors.NER(resourceName);
-  } else if (!DSUtils.isObject(params)) {
-    throw new DSErrors.IA('"params" must be an object!');
+  } else if (!DSUtils._o(params)) {
+    throw DSUtils._oErr('params');
   }
 
   definition.logFn('ejectAll', params, options);
 
-  var resource = _this.store[resourceName];
+  var resource = _this.s[resourceName];
   if (DSUtils.isEmpty(params)) {
     resource.completedQueries = {};
   }
   var queryHash = DSUtils.toJson(params);
-  var items = _this.filter(definition.name, params);
+  var items = _this.filter(definition.n, params);
   var ids = [];
   DSUtils.forEach(items, function (item) {
     if (item && item[definition.idAttribute]) {
@@ -28,7 +28,7 @@ function ejectAll(resourceName, params, options) {
   });
 
   DSUtils.forEach(ids, function (id) {
-    _this.eject(definition.name, id, options);
+    _this.eject(definition.n, id, options);
   });
 
   delete resource.completedQueries[queryHash];