+++ src/datastore/sync_methods/index.js
@@ -6,14 +6,14 @@ var R = DSErrors.R;
 
 function changes(resourceName, id, options) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
+  var definition = _this.defs[resourceName];
   options = options || {};
 
   id = DSUtils.resolveId(definition, id);
   if (!definition) {
     throw new NER(resourceName);
-  } else if (!DSUtils.isString(id) && !DSUtils.isNumber(id)) {
-    throw new IA('"id" must be a string or a number!');
+  } else if (!DSUtils._sn(id)) {
+    throw DSUtils._snErr('id');
   }
   options = DSUtils._(definition, options);
 
@@ -22,9 +22,9 @@ function changes(resourceName, id, options) {
   var item = _this.get(resourceName, id);
   if (item) {
     if (DSUtils.w) {
-      _this.store[resourceName].observers[id].deliver();
+      _this.s[resourceName].observers[id].deliver();
     }
-    var diff = DSUtils.diffObjectFromOldObject(item, _this.store[resourceName].previousAttributes[id], DSUtils.equals, options.ignoredChanges);
+    var diff = DSUtils.diffObjectFromOldObject(item, _this.s[resourceName].previousAttributes[id], DSUtils.equals, options.ignoredChanges);
     DSUtils.forOwn(diff, function (changeset, name) {
       var toKeep = [];
       DSUtils.forOwn(changeset, function (value, field) {
@@ -45,14 +45,14 @@ function changes(resourceName, id, options) {
 
 function changeHistory(resourceName, id) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
-  var resource = _this.store[resourceName];
+  var definition = _this.defs[resourceName];
+  var resource = _this.s[resourceName];
 
   id = DSUtils.resolveId(definition, id);
-  if (resourceName && !_this.definitions[resourceName]) {
+  if (resourceName && !_this.defs[resourceName]) {
     throw new NER(resourceName);
-  } else if (id && !DSUtils.isString(id) && !DSUtils.isNumber(id)) {
-    throw new IA('"id" must be a string or a number!');
+  } else if (id && !DSUtils._sn(id)) {
+    throw DSUtils._snErr('id');
   }
 
   definition.logFn('changeHistory', id);
@@ -73,14 +73,14 @@ function changeHistory(resourceName, id) {
 
 function compute(resourceName, instance) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
+  var definition = _this.defs[resourceName];
 
-  instance = DSUtils.resolveItem(_this.store[resourceName], instance);
+  instance = DSUtils.resolveItem(_this.s[resourceName], instance);
   if (!definition) {
     throw new NER(resourceName);
   } else if (!instance) {
     throw new R('Item not in the store!');
-  } else if (!DSUtils.isObject(instance) && !DSUtils.isString(instance) && !DSUtils.isNumber(instance)) {
+  } else if (!DSUtils._o(instance) && !DSUtils._sn(instance)) {
     throw new IA('"instance" must be an object, string or number!');
   }
 
@@ -94,7 +94,7 @@ function compute(resourceName, instance) {
 }
 
 function createInstance(resourceName, attrs, options) {
-  var definition = this.definitions[resourceName];
+  var definition = this.defs[resourceName];
   var item;
 
   attrs = attrs || {};
@@ -138,12 +138,12 @@ function digest() {
 
 function get(resourceName, id, options) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
+  var definition = _this.defs[resourceName];
 
   if (!definition) {
     throw new NER(resourceName);
-  } else if (!DSUtils.isString(id) && !DSUtils.isNumber(id)) {
-    throw new IA('"id" must be a string or a number!');
+  } else if (!DSUtils._sn(id)) {
+    throw DSUtils._snErr('id');
   }
 
   options = DSUtils._(definition, options);
@@ -151,7 +151,7 @@ function get(resourceName, id, options) {
   options.logFn('get', id, options);
 
   // cache miss, request resource from server
-  var item = _this.store[resourceName].index[id];
+  var item = _this.s[resourceName].index[id];
   if (!item && options.loadFromServer) {
     _this.find(resourceName, id, options);
   }
@@ -162,19 +162,19 @@ function get(resourceName, id, options) {
 
 function getAll(resourceName, ids) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
-  var resource = _this.store[resourceName];
+  var definition = _this.defs[resourceName];
+  var resource = _this.s[resourceName];
   var collection = [];
 
   if (!definition) {
     throw new NER(resourceName);
-  } else if (ids && !DSUtils.isArray(ids)) {
-    throw new IA('"ids" must be an array!');
+  } else if (ids && !DSUtils._a(ids)) {
+    throw DSUtils._aErr('ids');
   }
 
   definition.logFn('getAll', ids);
 
-  if (DSUtils.isArray(ids)) {
+  if (DSUtils._a(ids)) {
     var length = ids.length;
     for (var i = 0; i < length; i++) {
       if (resource.index[ids[i]]) {
@@ -190,14 +190,14 @@ function getAll(resourceName, ids) {
 
 function hasChanges(resourceName, id) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
+  var definition = _this.defs[resourceName];
 
   id = DSUtils.resolveId(definition, id);
 
   if (!definition) {
     throw new NER(resourceName);
-  } else if (!DSUtils.isString(id) && !DSUtils.isNumber(id)) {
-    throw new IA('"id" must be a string or a number!');
+  } else if (!DSUtils._sn(id)) {
+    throw DSUtils._snErr('id');
   }
 
   definition.logFn('hasChanges', id);
@@ -211,8 +211,8 @@ function hasChanges(resourceName, id) {
 }
 
 function lastModified(resourceName, id) {
-  var definition = this.definitions[resourceName];
-  var resource = this.store[resourceName];
+  var definition = this.defs[resourceName];
+  var resource = this.s[resourceName];
 
   id = DSUtils.resolveId(definition, id);
   if (!definition) {
@@ -231,8 +231,8 @@ function lastModified(resourceName, id) {
 }
 
 function lastSaved(resourceName, id) {
-  var definition = this.definitions[resourceName];
-  var resource = this.store[resourceName];
+  var definition = this.defs[resourceName];
+  var resource = this.s[resourceName];
 
   id = DSUtils.resolveId(definition, id);
   if (!definition) {
@@ -249,14 +249,14 @@ function lastSaved(resourceName, id) {
 
 function previous(resourceName, id) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
-  var resource = _this.store[resourceName];
+  var definition = _this.defs[resourceName];
+  var resource = _this.s[resourceName];
 
   id = DSUtils.resolveId(definition, id);
   if (!definition) {
     throw new NER(resourceName);
-  } else if (!DSUtils.isString(id) && !DSUtils.isNumber(id)) {
-    throw new IA('"id" must be a string or a number!');
+  } else if (!DSUtils._sn(id)) {
+    throw DSUtils._snErr('id');
   }
 
   definition.logFn('previous', id);