+++ src/datastore/sync_methods/linkInverse.js
@@ -3,17 +3,17 @@ var DSErrors = require('../../errors');
 
 function linkInverse(resourceName, id, relations) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
+  var definition = _this.defs[resourceName];
 
   relations = relations || [];
 
   id = DSUtils.resolveId(definition, id);
   if (!definition) {
     throw new DSErrors.NER(resourceName);
-  } else if (!DSUtils.isString(id) && !DSUtils.isNumber(id)) {
-    throw new DSErrors.IA('"id" must be a string or a number!');
-  } else if (!DSUtils.isArray(relations)) {
-    throw new DSErrors.IA('"relations" must be an array!');
+  } else if (!DSUtils._sn(id)) {
+    throw DSUtils._snErr('id');
+  } else if (!DSUtils._a(relations)) {
+    throw DSUtils._aErr('relations');
   }
 
   definition.logFn('linkInverse', id, relations);
@@ -21,14 +21,14 @@ function linkInverse(resourceName, id, relations) {
   var linked = _this.get(resourceName, id);
 
   if (linked) {
-    DSUtils.forOwn(_this.definitions, function (d) {
+    DSUtils.forOwn(_this.defs, function (d) {
       DSUtils.forOwn(d.relations, function (relatedModels) {
         DSUtils.forOwn(relatedModels, function (defs, relationName) {
-          if (relations.length && !DSUtils.contains(relations, d.name)) {
+          if (relations.length && !DSUtils.contains(relations, d.n)) {
             return;
           }
-          if (definition.name === relationName) {
-            _this.linkAll(d.name, {}, [definition.name]);
+          if (definition.n === relationName) {
+            _this.linkAll(d.n, {}, [definition.n]);
           }
         });
       });