+++ src/datastore/sync_methods/unlinkInverse.js
@@ -3,17 +3,17 @@ var DSErrors = require('../../errors');
 
 function unlinkInverse(resourceName, id, relations) {
   var _this = this;
-  var definition = _this.definitions[resourceName];
+  var definition = _this.defs[resourceName];
 
   relations = relations || [];
 
   id = DSUtils.resolveId(definition, id);
   if (!definition) {
     throw new DSErrors.NER(resourceName);
-  } else if (!DSUtils.isString(id) && !DSUtils.isNumber(id)) {
-    throw new DSErrors.IA('"id" must be a string or a number!');
-  } else if (!DSUtils.isArray(relations)) {
-    throw new DSErrors.IA('"relations" must be an array!');
+  } else if (!DSUtils._sn(id)) {
+    throw DSUtils._snErr('id');
+  } else if (!DSUtils._a(relations)) {
+    throw DSUtils._aErr('relations');
   }
 
   definition.logFn('unlinkInverse', id, relations);
@@ -21,12 +21,12 @@ function unlinkInverse(resourceName, id, relations) {
   var linked = _this.get(resourceName, id);
 
   if (linked) {
-    DSUtils.forOwn(_this.definitions, function (d) {
+    DSUtils.forOwn(_this.defs, function (d) {
       DSUtils.forOwn(d.relations, function (relatedModels) {
         DSUtils.forOwn(relatedModels, function (defs, relationName) {
-          if (definition.name === relationName) {
+          if (definition.n === relationName) {
             DSUtils.forEach(defs, function (def) {
-              DSUtils.forEach(_this.store[def.name].collection, function (item) {
+              DSUtils.forEach(_this.s[def.name].collection, function (item) {
                 if (def.type === 'hasMany' && item[def.localField]) {
                   var index;
                   DSUtils.forEach(item[def.localField], function (subItem, i) {